src/main/java/org/cp/elements/util/convert/support/CharacterConverter.java
/*
 * Copyright (c) 2011-Present. Codeprimate, LLC and authors.  All Rights Reserved.
 * <p/>
 * This software is licensed under the Codeprimate End User License Agreement (EULA).
 * This software is proprietary and confidential in addition to an intellectual asset
 * of the aforementioned authors.
 * <p/>
 * By using the software, the end-user implicitly consents to and agrees to be in compliance
 * with all terms and conditions of the EULA.  Failure to comply with the EULA will result in
 * the maximum penalties permissible by law.
 * <p/>
 * In short, this software may not be reverse engineered, reproduced, copied, modified
 * or distributed without prior authorization of the aforementioned authors, permissible
 * and expressed only in writing.  The authors grant the end-user non-exclusive, non-negotiable
 * and non-transferable use of the software "as is" without expressed or implied WARRANTIES,
 * EXTENSIONS or CONDITIONS of any kind.
 * <p/>
 * For further information on the software license, the end user is encouraged to read
 * the EULA @ ...
 */

package org.cp.elements.util.convert.support;

import org.cp.elements.util.convert.ConversionException;
import org.cp.elements.util.convert.ConverterAdapter;

/**
 * The CharacterConverter class converts an Object value into a Character.
 * <p/>
 * @author John J. Blum
 * @see java.lang.Character
 * @see java.lang.Object
 * @see org.cp.elements.util.convert.ConverterAdapter
 * @since 1.0.0
 */
@SuppressWarnings("unused")
public class CharacterConverter extends ConverterAdapter<Object, Character> {

  @Override
  public boolean canConvert(final Class<?> fromType, final Class<?> toType) {
    return (isAssignableTo(fromType, Character.class, String.class) && Character.class.equals(toType));
  }

  @Override
  public Character convert(final Object value) {
    if (value instanceof Character) {
      return (Character) value;
    }
    else if (value instanceof String) {
      final String valueString = value.toString();
      return (valueString.length() > 0 ? valueString.charAt(0) : '\0');
    }
    else {
      throw new ConversionException(String.format("The Object value (%1$s) is not a valid character!", value));
    }
  }

}
